Moved events and command management to own files. Not registered in this commit
