+++ src/delete-commands.ts
